make sure cargo test --all actually works
