+++ ci/test.sh
@@ -34,6 +34,6 @@ popd
 # Benches are checked in the `nightly` build
 # Format consistency is checked in the `beta` build
 printf "\n\n---- integration tests ----\n\n"
-cargo test --all --exclude sval_serde_no_std_tests --exclude sval_json_benches --exclude sval_fmt_tests
+cargo test --all --exclude sval_json_benches --exclude sval_fmt_tests
 
-cargo test -p sval_serde_no_std_tests
+cargo test -p sval_serde_no_alloc_tests
