+++ tests/serde_alloc/lib.rs
